% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Quarantine Shiny Application}
\usage{
launch_app()
}
\value{
This function does not return a value; it is called for its
side effect of launching a Shiny application.
}
\description{
This function starts the interactive Shiny app included in the package,
which allows users to explore the quarantine data.
}
